send ajax call to delete time slot only if it now a new one

+++ laravel/resources/js/components/Timeline/DateRangePicker.vue
@@ -38,7 +38,9 @@
             deleteItem() {
                 if (confirm('Delete this period?')) {
                     this.$emit('delete', this.period);
-                    axios.delete('/timeline/' + this.period.id);
+                    if (this.period.id > 0) {
+                        axios.delete('/timeline/' + this.period.id);
+                    }
                 }
             },
